+++ CHANGELOG.md
@@ -4,7 +4,9 @@ Changelog
 dev
 ---
 
-* -
+* Updated the used version of the underlying [libfuzzy
+  library](https://ssdeep-project.github.io/ssdeep/) to the current `master`
+  (last update: 2019-02-21).
 
 0.3.0 (2023-11-11)
 ------------------