+++ CHANGELOG.md
@@ -7,7 +7,8 @@ dev
 * Updated the used version of the underlying [libfuzzy
   library](https://ssdeep-project.github.io/ssdeep/) to the current `master`
   (last update: 2019-02-21).
-* The `configure` script is now executed directly instead of via `sh`.
+* The `configure` script is now executed directly instead of via `sh` when
+  building `libfuzzy`.
 * Improved the obtaining of the source directory when building `libfuzzy`.
 
 0.3.0 (2023-11-11)