+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## dev
 
+* -
+
+## 0.4.0 (2023-11-12)
+
 * Updated the used version of the underlying [libfuzzy
   library](https://ssdeep-project.github.io/ssdeep/) to the current `master`
   (last update: 2019-02-21).